requirements.txt
groq
streamlit==1.27.2  # currently pinned to this version
python-dotenv